再也不用git合并上游了，修的我烦了 Combat: - 基础战斗系统，召唤（SMN）满级循环：gauge、爆发、宝石兽/幻神召唤物 - 仇恨 / 敌对列表系统 - 基础伤害与治疗计算 - 敌人自动攻击与时间轴（timeline）机制
Duty / Director:
- 基础副本 director 系统（8Hz 时间引擎、AOE、可见性控制）
- 半成品伊弗利特歼灭战（w1fa_2，一阶段完成）
- 木人召唤 !dummy

Features:
- 支持染色
- 更完整的事件网络交互包（EventAction / EventFinish 拆分）

Fixes:
- 修复理发师（美容室）卡死黑屏
- 修复上游错误的 opcode / 包体大小
- 跨图与进出副本正确携带战斗状态，召唤物不再丢失、UI 不再失同步
- NPC 寻路重算后不再折返起点
- 大量技能 Lua 脚本（召唤 / 学者 / 通用）

+++ core/src/ipc/zone/server/enmity_list.rs
@@ -7,8 +7,9 @@ use crate::common::ObjectId;
 pub struct PlayerEnmity {
     /// The player's object ID.
     pub actor_id: ObjectId,
-    /// Value out of a 100, I think.
-    pub enmity: u32,
+    /// Enmity percentage, capped to 0..100.
+    #[brw(pad_after = 3)]
+    pub enmity: u8,
 }
 
 impl PlayerEnmity {
@@ -18,10 +19,11 @@ impl PlayerEnmity {
 #[binrw]
 #[derive(Debug, Clone, Default, PartialEq)]
 pub struct EnmityList {
-    pub count: u32,
-    #[brw(pad_after = 4)] // empty
+    #[brw(pad_after = 3)]
+    pub count: u8,
     #[br(count = count)]
     #[brw(pad_size_to = PlayerEnmity::SIZE * 8)]
+    #[brw(pad_after = 4)] // empty tail padding
     pub list: Vec<PlayerEnmity>,
 }
